# swagger.yaml
openapi: 3.0.0
info:
  title: library management api
  version: 1.0.0
  description: simple library management system with book reservations
servers:
  - url: http://localhost:8080/api/v1
    description: local dev server
components:
  schemas:
...
    ReservationResponse:
      type: object
      properties:
        book_id:
          type: string
          description: book id
        user_id:
          type: string
          description: user id
        message:
          type: string
          description: success message
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          description: success message
    Error:
      type: object
      properties:
        error:
          type: string
          description: error message
  securitySchemes:
    user_id:
      type: apiKey
      in: header
      name: user_id
      description: user id for authentication
paths:
...
  /books/{id}/reserve:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: book id
    post:
      summary: reserve a book
      security:
        - user_id: []
      responses:
        "200":
          description: book reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: book already reserved or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: book or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: cancel a reservation
      security:
        - user_id: []
      responses:
        "200":
          description: reservation canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: book not reserved by this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}/reservations:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: user id
    get:
      summary: get user’s reserved books
      security:
        - user_id: []
      responses:
        "200":
          description: list of reserved book ids
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "403":
          description: unauthorized access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"